Initial commit
Example_folder was created for writing a file organizer script, its function will be create new directories based on extension category.

For instance, if we have extensions like .docx, .txt, .xslx, and .ppt, we will be talking about Office or Text related files, so they will be stored in a new directory according to its category.

So far, we have a code that is able to create directories based on a dictionary named 'extensions', and also a function which can extract extensions .ext from a find shell command, check_files().

Our drawback lays on source for check_files() function, which has to be corrected for accepting os.listdir(files) and retrieving a list of total extensions.